requirements.txt
asgiref==3.2.10
attrs==20.1.0
boto3==1.14.55
botocore==1.17.55
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-rest-auth==0.9.5
django-storages==1.10
djangorestframework==3.11.1
docutils==0.15.2
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.11
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.25.10
wincertstore==0.2
